feat(front): add components module (global_components)
