+++ packages/front/src/modules/association/AssociationService.ts
@@ -10,7 +10,7 @@ import IdentifierHelper from "../../shared/helpers/IdentifierHelper";
 import ProviderValueHelper from "../../shared/helpers/ProviderValueHelper";
 
 export class AssociationService {
-    async getAssociation(id: string, user: User): Promise<{ type: "REDIRECT" | "SUCCESS" | "ERROR", data?: {association: AssociationDto, subventions: unknown, versements: DefaultObject<Versement[]> | null } }> {
+    async getAssociation(id: string, user: User): Promise<{ type: "REDIRECT" | "SUCCESS" | "ERROR", data?: {association: AssociationDto, subventions: unknown, versements: Versement[] } }> {
         const type = IdentifierHelper.findType(id);
         if (type === "UNKNOWN") return { type: "ERROR" }; // TODO send error
 
@@ -32,31 +32,15 @@ export class AssociationService {
         if (!association)  return { type: "ERROR" };
 
         const subventions = this.formatSubvention(association);
-        const versements = this.formatVersements(association.versements);
 
         return {
             type: "SUCCESS",
             data: {
                 association,
                 subventions,
-                versements,
+                versements: association.versements || [],
             }
         }
-    }
-
-    private formatVersements(versements?: Versement[]) {
-        if (!versements) return null;
-        const versementsByYear = versements.reduce((acc, versement) => {
-            const date = new Date(versement.dateOperation.value);
-            const year = date.getFullYear();
-            if (!acc[year]) acc[year] = [];
-            acc[year].push(versement);
-            return acc;
-        }, {} as DefaultObject<Versement[]>);
-
-        Object.values(versementsByYear).forEach(v => v.sort((a, b) => new Date(b.dateOperation.value).getTime() - new Date(a.dateOperation.value).getTime()))
-
-        return versementsByYear;
     }
 
     private formatSubvention(association: IAssociation) {
