+++ packages/front/src/modules/etablissement/EtablissementService.ts
@@ -20,7 +20,6 @@ export class EtablissementService {
             const etablissement = result.data.etablissement as IEtablissement;
             const association = etablissement.association;
             const subventions = this.formatSubvention(etablissement);
-            const versements = this.formatVersements(etablissement.versements);
 
             return {
                 type: "SUCCESS",
@@ -28,27 +27,12 @@ export class EtablissementService {
                     association,
                     etablissement,
                     subventions,
-                    versements
+                    versements: etablissement.versements || [],
                 }
             }
         }  catch (e) {
             return { type: "ERROR" }; // TODO send error
         }  
-    }
-
-    private formatVersements(versements?: Versement[]) {
-        if (!versements) return null;
-        const versementsByYear = versements.reduce((acc, versement) => {
-            const date = new Date(versement.dateOperation.value);
-            const year = date.getFullYear();
-            if (!acc[year]) acc[year] = [];
-            acc[year].push(versement);
-            return acc;
-        }, {} as DefaultObject<Versement[]>);
-
-        Object.values(versementsByYear).forEach(v => v.sort((a, b) => new Date(b.dateOperation.value).getTime() - new Date(a.dateOperation.value).getTime()))
-
-        return versementsByYear;
     }
 
     private formatSubvention(etablisement: IEtablissement) {
