+++ packages/front/src/modules/global_components/index.ts
@@ -1,0 +1,5 @@
+import * as versements from "./versements/index"
+
+export {
+    versements 
+}
